pyproject.toml
[tool.poetry]
name = "repl_python3_HypercuBot"
version = "0.1.0"
description = ""
authors = ["Rowan Fortier & Chris McElroy"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
Flask = "^2.0.2"
pil-compat = "^1.0.0"
cube = "^0.0.3"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
